Move inversion examples to package docstring
They were hidden in the misfit module. It's better to have them clear
where everyone will see.
